refactor: move debug structs to their own module
